refactor(FR-3228): extract SessionReclamationStatusPopover and reuse it in session detail idle checks
Move the reclamation popover (condition sentence, per-resource traffic-light
rows, level legend) and the color helpers (getUtilizationCheckerColor,
getOverallReclamation, useReclamationColorMap) into a shared presentational
component taking plain resources/operator props — a Relay fragment component
would not fit SessionIdleChecks' legacy compute_session query path, and both
call sites already hold the parsed idle_checks extra. The session detail's old
dark tooltip ('utilization / threshold' with colored values) is replaced by the
same popover used in the session list column.
